requirements.txt
# Core LangChain dependencies
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-google-genai

# Vector database and embeddings
chromadb
sentence-transformers

# Data processing
pandas
numpy

# Environment and configuration
python-dotenv

# Streamlit web interface
streamlit

# Text processing and search
rank-bm25
nltk

# Machine Learning
# torch
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
typing-extensions

# Google AI integration
google-genai

